feat(#44): 검색과 프로필 정보에서 사용 할 공용 컴포넌트 분리
